requirements.txt
pyyaml
python-liquid
munch
htmlmin
